Refactor main.js: torrent controller
